feat(condo): DOMA-11548 rename marketingConsent to hasMarketingConsent
